+++ ricaun.Revit.UI.Tests/Resources/ResourceTests.cs
@@ -6,7 +6,6 @@ namespace ricaun.Revit.UI.Tests.Resources
     public class ResourceTests
     {
         const string ResourceNameIco = "Resources/Images/Revit.ico";
-        const string ResourceNamePng = "Resources/Images/Revit32.png";
 
         [Test]
         public void GetBitmapSource_NotNull()
@@ -14,14 +13,6 @@ namespace ricaun.Revit.UI.Tests.Resources
             Console.WriteLine(ResourceNameIco.GetBitmapSource());
             Assert.IsNotNull(ResourceNameIco.GetBitmapSource());
             Assert.IsNotNull(("/" + ResourceNameIco).GetBitmapSource());
-        }
-
-        [Test]
-        public void GetBitmapSource_NotNull_Png()
-        {
-            Console.WriteLine(ResourceNamePng.GetBitmapSource());
-            Assert.IsNotNull(ResourceNamePng.GetBitmapSource());
-            Assert.IsNotNull(("/" + ResourceNamePng).GetBitmapSource());
         }
 
         [Test]
@@ -38,6 +29,25 @@ namespace ricaun.Revit.UI.Tests.Resources
             var component = "NotExist/" + ResourceNameIco;
             var source = component.GetBitmapSource();
             Assert.IsNull(source);
+        }
+
+        [TestCase(32)]
+        public void GetBitmapSource_Width(int width)
+        {
+            var source = ResourceNameIco.GetBitmapSource();
+            Assert.AreEqual(width, source.Width);
+        }
+
+        [TestCase(0, 16)]
+        [TestCase(10, 10)]
+        [TestCase(16, 16)]
+        [TestCase(32, 32)]
+        [TestCase(64, 32)]
+        [TestCase(128, 32)]
+        public void GetBitmapFrame_Width(int width, int expected)
+        {
+            var source = ResourceNameIco.GetBitmapSource().GetBitmapFrame(width);
+            Assert.AreEqual(expected, source.Width);
         }
     }
 }